classification edited

+++ Facility/Server/src/classification/repositories/classification.repository.ts
@@ -12,10 +12,11 @@ import { ClassificationNotFountException } from 'src/common/notFoundExceptions/n
 import { assignDtoPropToEntity, createDynamicCyperCreateQuery, createDynamicCyperObject } from 'src/common/func/neo4js.func';
 
 import { CustomNeo4jError, Neo4jService } from 'src/sgnm-neo4j/src';
-import { newError } from 'neo4j-driver-core';
+
 import { create_node__must_entered_error, create_node__node_not_created_error, find_by_realm_with_tree_structure__not_entered_error, find_by_realm__not_entered_error, find_by_realm__not_found_error, find_with_children_by_realm_as_tree_error, find_with_children_by_realm_as_tree__find_by_realm_error, find_with_children_by_realm_as_tree__not_entered_error, tree_structure_not_found_by_realm_name_error } from 'src/sgnm-neo4j/src/constant/custom.error.object';
+import { GeciciInterface } from 'src/common/interface/gecici.interface';
 @Injectable()
-export class ClassificationRepository implements BaseGraphDatabaseInterfaceRepository<Classification> {
+export class ClassificationRepository implements GeciciInterface<Classification> {
   constructor(private readonly neo4jService: Neo4jService) {}
 
   async findOneById(id: string) {
@@ -116,11 +117,11 @@ export class ClassificationRepository implements BaseGraphDatabaseInterfaceRepos
     let classification = new Classification();
     let classificationObject = assignDtoPropToEntity(classification, createClassificationDto);
     let value;
-    if (classificationObject['labels'].length > 0) {
-      value = await this.createNode(classificationObject['entity'], classificationObject['labels']);
+    if (classificationObject['labels']) {
+      value = await this.neo4jService.createNode(classificationObject['entity'], classificationObject['labels']);
     }
     else {
-      value = await this.createNode(classificationObject['entity']);
+      value = await this.neo4jService.createNode(classificationObject['entity']);
     }
     value['properties']['id'] = value['identity'].low;
     const result = {id:value['identity'].low, labels: value['labels'], properties: value['properties']}
@@ -145,24 +146,32 @@ export class ClassificationRepository implements BaseGraphDatabaseInterfaceRepos
   }
   async delete(_id: string) {
     try {
-      const deletedNode = await this.neo4jService.delete(_id);
-      if (!deletedNode) {
-        throw new ClassificationNotFountException(_id);
+     
+      let hasParent = await this.neo4jService.findParentById(_id);
+      let deletedNode;
+      if (hasParent['records'].length > 0) {
+         let hasChildren =  await this.neo4jService.findChildrenById(_id);       
+         if (hasChildren['records'].length == 0) {
+           deletedNode = await this.neo4jService.delete(_id);
+           if (!deletedNode) {
+             throw new ClassificationNotFountException(_id);
+          }
+         }
       }
       return deletedNode;
     } catch (error) {
-      const { code } = error.response;
+      const { code, message } = error.response;
       if (code === CustomNeo4jError.HAS_CHILDREN) {
         nodeHasChildException(_id);
       } else {
-        throw new HttpException(error, HttpStatus.INTERNAL_SERVER_ERROR);
+        throw new HttpException(message, code);
       }
     }
   }
   async changeNodeBranch(_id: string, target_parent_id: string) {
     try {
-      await this.deleteRelations(_id);
-      await this.addRelations(_id, target_parent_id);
+      await this.neo4jService.deleteRelations(_id);
+      await this.neo4jService.addRelations(_id, target_parent_id);
     } catch (error) {
       throw new HttpException(error, HttpStatus.INTERNAL_SERVER_ERROR);
     }
@@ -184,38 +193,14 @@ export class ClassificationRepository implements BaseGraphDatabaseInterfaceRepos
       if (!node) {
         throw new ClassificationNotFountException(key);
       }
-      return node;
+      const result = {id:node['identity'].low, labels: node['labels'], properties: node['properties']}
+      return result;
     } catch (error) {
       throw new HttpException(error, HttpStatus.INTERNAL_SERVER_ERROR);
     }
   }
 
-//////////neo4j
-async createNode(
-  params: object,
-  labels?: string[],
-  ) {
-  try {
-    if(!params ){
-      throw new HttpException(create_node__must_entered_error,400);
-    }
-    const cyperQuery = createDynamicCyperCreateQuery(params,labels);
-    const res = await this.neo4jService.write(cyperQuery, params);
-    if(!res["records"][0].length){
-      throw new HttpException(create_node__node_not_created_error,400);
-    }
-    return res["records"][0]["_fields"][0];
-  } catch (error) {
-    if (error.response.code) {
-      throw new HttpException(
-        { message: error.response.message, code: error.response.code },
-        error.status
-      );
-    }else {
-      throw newError(error, "500");
-    }
-  }
-}
+
 
 }
